// front-end/src/components/userNavBar/userNavBar.interface.ts
/*
 * @Author: sunji 2025506282@qq.com
 * @Date: 2022-09-30 16:00:51
 * @LastEditors: sunji 2025506282@qq.com
 * @LastEditTime: 2022-10-09 13:51:09
 * @FilePath: \front-end\src\components\tabBar\tabBar.interface.ts
 * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
 */
export interface IMenu {
  label: string
  value: string
  isSplit?: boolean
  children?: IMenu[]
}